utrans0* fails with multiple labels cc @TiarkRompf
The problematic case is stp/sel2 - stp/llsel2 :
T1 <:* x.L1 <:n x.L2
we can still get T1 <:(_+1) x.L1
but that doesn't give us T<:(_+1) x.L2 since labels are different.

So it seems that the current rules in stpl are not sufficient when we
have multiple labels.