sdc_typ <: sdc_typ trans case.
if trans_on is only defined for stp true, then we actually need to
change stp_transf to take both false sides. for this case so far,
we could have proved the false case inductively if it were true.
getting confused as to where it's supposed to matter.

cc @TiarkRompf