Fix remaining admits.
To do the mutual induction between stp2 and val_type for extending the
store, I changed the existential on stp2 in val_type definitions to a
top-level universal, which is semantically equivalent. Some proofs
required minor changes, but for the mutual induction, Coq picks up the
IH on stp2 this way, while it didn't with the existential.

cc @TiarkRompf