Done with multiple members.
cc @TiarkRompf

only admit is for extending GH, but we know how to do that with
splicing, etc. it seems necessary to extend GH for subst, once we avoid
vtp in has_type.

next step is to confirm that the non-alternative development (without
vtp, but with trans pushback) still works.