go back to pty which only allows variables (but not values) at head of path, and reintroduce the strong stp_proj rules based on small-step reduction
A tempting alternative would have been to keep the pty which allows values at head of path, so that substitution is simpler, and to make vtp_widen work, have an intermediate step where stp is transformed into tight-bounds stp, as in lfdot. But then, tight-bounds are not enforced any more when the env grows (we must define it that way), so substitution (which shrinks/empties the env) would have to convert weak stp_proj nodes into strong ones, which is exactly what we wanted to avoid, but, as we see, we cannot avoid it.